Changed json url

+++ app/src/main/java/net/ducko/geesemodpack/ForgeDownloadTask.java
@@ -4,7 +4,6 @@ import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.net.URL;
 
 public class ForgeDownloadTask extends Task {
@@ -19,27 +18,25 @@ public class ForgeDownloadTask extends Task {
 		this.version = version;
 		this.filename = "forge-" + version + "-installer.jar";
 		this.filepath = path;
-		/*ClassLoader loader = Thread.currentThread().getContextClassLoader();
-		URL url = loader.getResource("files");
-		String localPath = url.getPath();
-		this.filepath =  localPath + "/";*/
 	}
 	
 	@Override
 	public int execute() {
 		
 		final String url = forgeURL + version + "/" + filename;
-		File file = new File(filepath, filename);
+		File file = new File(filepath + filename);
 		
-		try (BufferedInputStream in = new BufferedInputStream(new URL(url).openStream());
-				  FileOutputStream fileOutputStream = new FileOutputStream(file)) {
-		    byte dataBuffer[] = new byte[1024];
-		    int bytesRead;
-		    while ((bytesRead = in.read(dataBuffer, 0, 1024)) != -1) {
-		        fileOutputStream.write(dataBuffer, 0, bytesRead);
-		    }
-		} catch (IOException e) {
-			e.printStackTrace();
+		if (!file.exists()) {
+			try (BufferedInputStream in = new BufferedInputStream(new URL(url).openStream());
+					  FileOutputStream fileOutputStream = new FileOutputStream(file)) {
+			    byte dataBuffer[] = new byte[1024];
+			    int bytesRead;
+			    while ((bytesRead = in.read(dataBuffer, 0, 1024)) != -1) {
+			        fileOutputStream.write(dataBuffer, 0, bytesRead);
+			    }
+			} catch (IOException e) {
+				e.printStackTrace();
+			}
 		}
 		
 		return weight;
